refactor: split starmap view owners
